Update shipping zones and warehouses cards in chaannel page
