+++ src/channels/components/ShippingZones/index.tsx
@@ -1,0 +1,2 @@
+export * from "./ShippingZones";
+export { default } from "./ShippingZones";
