--- src/channels/components/ShippingZonesCard/index.tsx
@@ -1,2 +1,0 @@
-export * from "./ShippingZonesCard";
-export { default } from "./ShippingZonesCard";
