+++ src/channels/components/Warehouses/index.tsx
@@ -1,0 +1,2 @@
+export * from "./Warehouses";
+export { default } from "./Warehouses";
