--- src/channels/components/WarehousesCard/index.tsx
@@ -1,2 +1,0 @@
-export * from "./WarehousesCard";
-export { default } from "./WarehousesCard";
